requirements.txt
requests==2.32.3
python-dotenv==1.0.1
kafka-python==2.2.15
pyspark==4.0.1
boto3==1.37.38